@ custom_modules fuse hamming.cpp file back into custom.cpp.
