+++ PhysiCell/custom_modules/custom.h
@@ -72,7 +72,6 @@ using namespace BioFVM;
 using namespace PhysiCell;
 
 
-
 // setup functions to help us along
 
 void create_cell_types( void );
@@ -108,18 +107,8 @@ void bfollicular_cell_phenotype( Cell* pCell, Phenotype& phenotype , double dt )
 //void bmemory_cell_phenotype( Cell* pCell, Phenotype& phenotype , double dt );
 //void antibody_phenotype( Cell* pCell, Phenotype& phenotype , double dt );
 
-// custom constantes
-static const std::string B_NAIVE_NAME {"B_naive"}; //from config XML
-
-//letters not in the human amino acid alphabet: b,j,o,u,x,z
-//static const double ALPHABET[] {'a','c','d','e','f','g','h','i','k','l','m','n','p','q','r','s','t','v','w','y'};  // humman amino acide alphabet
-static const double ALPHABET[] {'a','t','c','g'};  // oligo nucleotide alphabet
-static const double PAD {0};
-
-// LEN_VECTOR_SEQUENCE >= max(LEN_ANTIGEN_SEQUENCE, LEN_ANTIBODY_SEQUENCE) >= min(LEN_ANTIGEN_SEQUENCE, LEN_ANTIBODY_SEQUENCE) >= LEN_AMINOCOMPLETE
-static const int LEN_VECTOR_SEQUENCE = 16;
-static const int LEN_ANTIBODY_SEQUENCE = 8;
-static const int LEN_ANTIGEN_SEQUENCE = 8;
-static const int LEN_AMINOCOMPLETE = 5;  // number of matching antigen antibody amino sequences that account for 100% affinity.
-static const int MUTATION = 1;  // number antibody sequence mutations per follicular B cell division.
-static const std::vector<double> EMPTY_VECTOR(LEN_VECTOR_SEQUENCE, PAD);  // generate empty antigen antybody vector.
+//std::vector<double> get_vector_variable( Cell* pCell, std::string name );
+void printSequence( std::vector<double>& sequence );
+std::vector<double> generateSequence( int lenSequence );
+void mutateSequence( std::vector<double>& sequence, int mutations );
+double alignment( Vector_Variable antigenSequence, Vector_Variable antibodySequence );
